Refactor: Modular directory structure to avoid large files
- Refactored 4 large files (3061 lines) into 19 focused modules
- Added comprehensive module-level documentation in mod.rs files
- Updated README.md with module structure section

Modules refactored:
- facilitator/ (1109 lines → 3 files: mod, coinbase, tests)
- crypto/ (661 lines → 5 files: mod, jwt, eip712, signature, tests)
- types/ (542 lines → 6 files: mod, network, payment, facilitator, discovery, constants)
- middleware/ (749 lines → 5 files: mod, config, payment, service, tests)

Benefits:
- Better code organization with single responsibility principle
- Improved maintainability with smaller, focused files
- Enhanced documentation with examples in each module
- Reduced merge conflicts for team collaboration
- Follows Rust best practices for module documentation
