--- src/crypto.rs
@@ -1,661 +1,0 @@
-//! Cryptographic utilities for x402 payments
-
-use crate::{Result, X402Error};
-use ethereum_types::{Address, H256, U256};
-use k256::ecdsa::{RecoveryId, Signature as K256Signature};
-use secp256k1::{Message, Secp256k1, SecretKey};
-use serde_json::json;
-use std::str::FromStr;
-
-/// EIP-712 domain separator for EIP-3009 transfers
-pub const EIP712_DOMAIN: &str = r#"{"name":"USD Coin","version":"2","chainId":8453,"verifyingContract":"0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913"}"#;
-
-/// JWT utilities for authentication
-pub mod jwt {
-    use super::*;
-    use jsonwebtoken::{Algorithm, Header};
-
-    /// JWT claims for Coinbase API authentication
-    #[derive(Debug, serde::Serialize)]
-    struct Claims {
-        iss: String,
-        sub: String,
-        aud: String,
-        iat: u64,
-        exp: u64,
-        uri: String,
-    }
-
-    /// JWT options for authentication
-    #[derive(Debug, Clone)]
-    pub struct JwtOptions {
-        pub key_id: String,
-        pub key_secret: String,
-        pub request_method: String,
-        pub request_host: String,
-        pub request_path: String,
-    }
-
-    impl JwtOptions {
-        /// Create new JWT options
-        pub fn new(
-            key_id: impl Into<String>,
-            key_secret: impl Into<String>,
-            request_method: impl Into<String>,
-            request_host: impl Into<String>,
-            request_path: impl Into<String>,
-        ) -> Self {
-            Self {
-                key_id: key_id.into(),
-                key_secret: key_secret.into(),
-                request_method: request_method.into(),
-                request_host: request_host.into(),
-                request_path: request_path.into(),
-            }
-        }
-    }
-
-    /// Generate JWT token for Coinbase API authentication
-    pub fn generate_jwt(options: JwtOptions) -> Result<String> {
-        // Remove https:// if present
-        let request_host = options.request_host.trim_start_matches("https://");
-
-        let now = chrono::Utc::now().timestamp() as u64;
-        let exp = now + 300; // 5 minutes
-
-        let claims = Claims {
-            iss: options.key_id.clone(),
-            sub: options.key_id,
-            aud: request_host.to_string(),
-            iat: now,
-            exp,
-            uri: options.request_path,
-        };
-
-        let header = Header::new(Algorithm::HS256);
-        let key = jsonwebtoken::EncodingKey::from_secret(options.key_secret.as_bytes());
-        let token = jsonwebtoken::encode(&header, &claims, &key)
-            .map_err(|e| X402Error::config(format!("JWT encoding failed: {}", e)))?;
-
-        Ok(token)
-    }
-
-    /// Create an authorization header for Coinbase API requests
-    pub fn create_auth_header(
-        api_key_id: &str,
-        api_key_secret: &str,
-        request_host: &str,
-        request_path: &str,
-    ) -> Result<String> {
-        let options = JwtOptions::new(
-            api_key_id,
-            api_key_secret,
-            "POST", // Default to POST method
-            request_host,
-            request_path,
-        );
-
-        let token = generate_jwt(options)?;
-        Ok(format!("Bearer {}", token))
-    }
-
-    /// Create auth header with custom method
-    pub fn create_auth_header_with_method(
-        api_key_id: &str,
-        api_key_secret: &str,
-        request_method: &str,
-        request_host: &str,
-        request_path: &str,
-    ) -> Result<String> {
-        let options = JwtOptions::new(
-            api_key_id,
-            api_key_secret,
-            request_method,
-            request_host,
-            request_path,
-        );
-
-        let token = generate_jwt(options)?;
-        Ok(format!("Bearer {}", token))
-    }
-}
-
-/// EIP-712 typed data utilities
-pub mod eip712 {
-    use super::*;
-
-    /// EIP-712 domain separator
-    #[derive(Debug, Clone)]
-    pub struct Domain {
-        pub name: String,
-        pub version: String,
-        pub chain_id: u64,
-        pub verifying_contract: Address,
-    }
-
-    /// EIP-712 typed data structure
-    #[derive(Debug, Clone)]
-    pub struct TypedData {
-        pub domain: Domain,
-        pub primary_type: String,
-        pub types: serde_json::Value,
-        pub message: serde_json::Value,
-    }
-
-    /// Create EIP-712 hash for EIP-3009 transfer with authorization
-    pub fn create_transfer_with_authorization_hash(
-        domain: &Domain,
-        from: Address,
-        to: Address,
-        value: U256,
-        valid_after: U256,
-        valid_before: U256,
-        nonce: H256,
-    ) -> Result<H256> {
-        let types = json!({
-            "EIP712Domain": [
-                {"name": "name", "type": "string"},
-                {"name": "version", "type": "string"},
-                {"name": "chainId", "type": "uint256"},
-                {"name": "verifyingContract", "type": "address"}
-            ],
-            "TransferWithAuthorization": [
-                {"name": "from", "type": "address"},
-                {"name": "to", "type": "address"},
-                {"name": "value", "type": "uint256"},
-                {"name": "validAfter", "type": "uint256"},
-                {"name": "validBefore", "type": "uint256"},
-                {"name": "nonce", "type": "bytes32"}
-            ]
-        });
-
-        let message = json!({
-            "from": format!("{:?}", from),
-            "to": format!("{:?}", to),
-            "value": format!("0x{:x}", value),
-            "validAfter": format!("0x{:x}", valid_after),
-            "validBefore": format!("0x{:x}", valid_before),
-            "nonce": format!("{:?}", nonce)
-        });
-
-        let typed_data = TypedData {
-            domain: domain.clone(),
-            primary_type: "TransferWithAuthorization".to_string(),
-            types,
-            message,
-        };
-
-        hash_typed_data(&typed_data)
-    }
-
-    /// Hash EIP-712 typed data
-    pub fn hash_typed_data(typed_data: &TypedData) -> Result<H256> {
-        // Full EIP-712 implementation following the specification
-
-        let domain_separator = hash_domain(&typed_data.domain)?;
-        let struct_hash = hash_struct(
-            &typed_data.primary_type,
-            &typed_data.types,
-            &typed_data.message,
-        )?;
-
-        // EIP-712: hash(0x1901 || domain_separator || struct_hash)
-        let mut data = Vec::new();
-        data.extend_from_slice(&[0x19, 0x01]); // EIP-712 prefix
-        data.extend_from_slice(domain_separator.as_bytes());
-        data.extend_from_slice(struct_hash.as_bytes());
-
-        Ok(H256::from_slice(&keccak256(&data)))
-    }
-
-    /// Hash the domain separator
-    fn hash_domain(domain: &Domain) -> Result<H256> {
-        let domain_type_hash = keccak256(
-            b"EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)",
-        );
-
-        let name_hash = keccak256(domain.name.as_bytes());
-        let version_hash = keccak256(domain.version.as_bytes());
-        let chain_id_hash = keccak256(&domain.chain_id.to_be_bytes());
-        let verifying_contract_hash = keccak256(domain.verifying_contract.as_bytes());
-
-        let mut data = Vec::new();
-        data.extend_from_slice(&domain_type_hash);
-        data.extend_from_slice(&name_hash);
-        data.extend_from_slice(&version_hash);
-        data.extend_from_slice(&chain_id_hash);
-        data.extend_from_slice(&verifying_contract_hash);
-
-        Ok(H256::from_slice(&keccak256(&data)))
-    }
-
-    /// Hash a struct according to EIP-712
-    fn hash_struct(
-        primary_type: &str,
-        _types: &serde_json::Value,
-        message: &serde_json::Value,
-    ) -> Result<H256> {
-        // Full EIP-712 struct hashing implementation
-
-        // For TransferWithAuthorization, create the proper type hash
-        let type_hash = keccak256(
-            format!("{}(address from,address to,uint256 value,uint256 validAfter,uint256 validBefore,bytes32 nonce)", primary_type)
-            .as_bytes()
-        );
-
-        // Encode the message fields in the correct order
-        let encoded_message = encode_message_fields(message)?;
-        let message_hash = keccak256(&encoded_message);
-
-        // Combine type hash and message hash
-        let mut data = Vec::new();
-        data.extend_from_slice(&type_hash);
-        data.extend_from_slice(&message_hash);
-
-        Ok(H256::from_slice(&keccak256(&data)))
-    }
-
-    /// Encode message fields for hashing
-    fn encode_message_fields(message: &serde_json::Value) -> Result<Vec<u8>> {
-        // For TransferWithAuthorization, encode fields in the correct order
-        let mut encoded = Vec::new();
-
-        // Encode 'from' address (32 bytes, padded)
-        if let Some(from) = message.get("from") {
-            if let Some(addr_str) = from.as_str() {
-                let addr = Address::from_str(addr_str)
-                    .map_err(|_| X402Error::invalid_authorization("Invalid from address"))?;
-                let mut padded = [0u8; 32];
-                padded[12..32].copy_from_slice(addr.as_bytes());
-                encoded.extend_from_slice(&padded);
-            }
-        }
-
-        // Encode 'to' address (32 bytes, padded)
-        if let Some(to) = message.get("to") {
-            if let Some(addr_str) = to.as_str() {
-                let addr = Address::from_str(addr_str)
-                    .map_err(|_| X402Error::invalid_authorization("Invalid to address"))?;
-                let mut padded = [0u8; 32];
-                padded[12..32].copy_from_slice(addr.as_bytes());
-                encoded.extend_from_slice(&padded);
-            }
-        }
-
-        // Encode 'value' (32 bytes, big-endian)
-        if let Some(value) = message.get("value") {
-            if let Some(value_str) = value.as_str() {
-                let value_hex = value_str.trim_start_matches("0x");
-                let value_bytes = hex::decode(value_hex)
-                    .map_err(|_| X402Error::invalid_authorization("Invalid value format"))?;
-                let mut padded = [0u8; 32];
-                let start = 32 - value_bytes.len();
-                padded[start..].copy_from_slice(&value_bytes);
-                encoded.extend_from_slice(&padded);
-            }
-        }
-
-        // Encode 'validAfter' (32 bytes, big-endian)
-        if let Some(valid_after) = message.get("validAfter") {
-            if let Some(valid_after_str) = valid_after.as_str() {
-                let valid_after_hex = valid_after_str.trim_start_matches("0x");
-                let valid_after_bytes = hex::decode(valid_after_hex)
-                    .map_err(|_| X402Error::invalid_authorization("Invalid validAfter format"))?;
-                let mut padded = [0u8; 32];
-                let start = 32 - valid_after_bytes.len();
-                padded[start..].copy_from_slice(&valid_after_bytes);
-                encoded.extend_from_slice(&padded);
-            }
-        }
-
-        // Encode 'validBefore' (32 bytes, big-endian)
-        if let Some(valid_before) = message.get("validBefore") {
-            if let Some(valid_before_str) = valid_before.as_str() {
-                let valid_before_hex = valid_before_str.trim_start_matches("0x");
-                let valid_before_bytes = hex::decode(valid_before_hex)
-                    .map_err(|_| X402Error::invalid_authorization("Invalid validBefore format"))?;
-                let mut padded = [0u8; 32];
-                let start = 32 - valid_before_bytes.len();
-                padded[start..].copy_from_slice(&valid_before_bytes);
-                encoded.extend_from_slice(&padded);
-            }
-        }
-
-        // Encode 'nonce' (32 bytes)
-        if let Some(nonce) = message.get("nonce") {
-            if let Some(nonce_str) = nonce.as_str() {
-                let nonce_hex = nonce_str.trim_start_matches("0x");
-                let nonce_bytes = hex::decode(nonce_hex)
-                    .map_err(|_| X402Error::invalid_authorization("Invalid nonce format"))?;
-                if nonce_bytes.len() != 32 {
-                    return Err(X402Error::invalid_authorization("Nonce must be 32 bytes"));
-                }
-                encoded.extend_from_slice(&nonce_bytes);
-            }
-        }
-
-        Ok(encoded)
-    }
-
-    /// Keccak-256 hash function
-    fn keccak256(data: &[u8]) -> [u8; 32] {
-        use sha3::{Digest, Keccak256};
-        Keccak256::digest(data).into()
-    }
-
-    /// SHA3-256 hash function
-    ///
-    /// This function is available for applications that need SHA3-256 hashing
-    /// in addition to Keccak-256. While EIP-712 primarily uses Keccak-256,
-    /// SHA3-256 may be needed for other cryptographic operations.
-    pub fn sha3_256(data: &[u8]) -> [u8; 32] {
-        use sha3::{Digest, Sha3_256};
-        Sha3_256::digest(data).into()
-    }
-}
-
-/// Signature utilities
-pub mod signature {
-    use super::*;
-    use k256::ecdsa::VerifyingKey;
-
-    /// Verify an EIP-712 signature
-    pub fn verify_eip712_signature(
-        signature: &str,
-        message_hash: H256,
-        expected_address: Address,
-    ) -> Result<bool> {
-        let sig_bytes = hex::decode(signature.trim_start_matches("0x"))
-            .map_err(|_| X402Error::invalid_signature("Invalid hex signature"))?;
-
-        if sig_bytes.len() != 65 {
-            return Err(X402Error::invalid_signature("Signature must be 65 bytes"));
-        }
-
-        let r = H256::from_slice(&sig_bytes[0..32]);
-        let s = H256::from_slice(&sig_bytes[32..64]);
-        let v = sig_bytes[64];
-
-        let recovery_id = RecoveryId::try_from(v)
-            .map_err(|_| X402Error::invalid_signature("Invalid recovery ID"))?;
-
-        // Create k256 signature from r and s
-        let mut sig_bytes = [0u8; 64];
-        sig_bytes[0..32].copy_from_slice(r.as_bytes());
-        sig_bytes[32..64].copy_from_slice(s.as_bytes());
-
-        let k256_sig = K256Signature::try_from(&sig_bytes[..])
-            .map_err(|_| X402Error::invalid_signature("Invalid signature format"))?;
-
-        // Recover the public key
-        let verifying_key =
-            VerifyingKey::recover_from_prehash(message_hash.as_bytes(), &k256_sig, recovery_id)
-                .map_err(|_| X402Error::invalid_signature("Failed to recover public key"))?;
-
-        // Convert to Ethereum address
-        let recovered_address = ethereum_address_from_pubkey(&verifying_key)?;
-
-        Ok(recovered_address == expected_address)
-    }
-
-    /// Sign a message hash with a private key
-    pub fn sign_message_hash(message_hash: H256, private_key: &str) -> Result<String> {
-        let private_key_bytes = hex::decode(private_key.trim_start_matches("0x"))
-            .map_err(|_| X402Error::invalid_signature("Invalid hex private key"))?;
-
-        let secret_key = SecretKey::from_slice(&private_key_bytes)
-            .map_err(|_| X402Error::invalid_signature("Invalid private key"))?;
-
-        let secp = Secp256k1::new();
-        let message = Message::from_digest_slice(message_hash.as_bytes())
-            .map_err(|_| X402Error::invalid_signature("Invalid message hash"))?;
-
-        let signature = secp.sign_ecdsa(&message, &secret_key);
-        let serialized = signature.serialize_compact();
-
-        // Compute the recovery ID properly
-        // The recovery ID is used to recover the public key from the signature
-        let recovery_id = compute_recovery_id(&signature, &message, &secret_key)?;
-
-        // Convert to k256 signature for consistency
-        let _k256_sig = K256Signature::try_from(&serialized[..])
-            .map_err(|_| X402Error::invalid_signature("Failed to convert signature"))?;
-
-        // Create the full signature with recovery ID
-        let mut sig_bytes = [0u8; 65];
-        sig_bytes[0..32].copy_from_slice(&serialized[0..32]);
-        sig_bytes[32..64].copy_from_slice(&serialized[32..64]);
-        sig_bytes[64] = recovery_id;
-
-        Ok(format!("0x{}", hex::encode(sig_bytes)))
-    }
-
-    /// Convert a public key to an Ethereum address
-    fn ethereum_address_from_pubkey(pubkey: &k256::ecdsa::VerifyingKey) -> Result<Address> {
-        let pubkey_bytes = pubkey.to_sec1_bytes();
-        if pubkey_bytes.len() != 65 {
-            return Err(X402Error::invalid_signature("Invalid public key length"));
-        }
-
-        // Remove the first byte (0x04) and hash the remaining 64 bytes
-        let pubkey_hash = keccak256(&pubkey_bytes[1..]);
-
-        // Take the last 20 bytes as the address
-        let mut address_bytes = [0u8; 20];
-        address_bytes.copy_from_slice(&pubkey_hash[12..]);
-
-        Ok(Address::from(address_bytes))
-    }
-
-    /// Compute the recovery ID for a signature
-    fn compute_recovery_id(
-        signature: &secp256k1::ecdsa::Signature,
-        message: &Message,
-        private_key: &SecretKey,
-    ) -> Result<u8> {
-        let secp = Secp256k1::new();
-
-        // Get the public key from the private key
-        let public_key = private_key.public_key(&secp);
-
-        // Try both possible recovery IDs (0 and 1)
-        for recovery_id in 0..2 {
-            // Create RecoveryId from i32 (secp256k1 uses i32, not u8)
-            let recovery_id_enum = secp256k1::ecdsa::RecoveryId::from_i32(recovery_id as i32);
-            if let Ok(recovery_id_enum) = recovery_id_enum {
-                // Create a recoverable signature with this recovery ID
-                if let Ok(recoverable_sig) = secp256k1::ecdsa::RecoverableSignature::from_compact(
-                    &signature.serialize_compact(),
-                    recovery_id_enum,
-                ) {
-                    // Try to recover the public key using this recovery ID
-                    if let Ok(recovered_key) = secp.recover_ecdsa(message, &recoverable_sig) {
-                        // If the recovered key matches our public key, this is the correct recovery ID
-                        if recovered_key == public_key {
-                            return Ok(recovery_id);
-                        }
-                    }
-                }
-            }
-        }
-
-        Err(X402Error::invalid_signature(
-            "Could not determine recovery ID",
-        ))
-    }
-
-    /// Keccak-256 hash function
-    fn keccak256(data: &[u8]) -> [u8; 32] {
-        use sha3::{Digest, Keccak256};
-        Keccak256::digest(data).into()
-    }
-
-    /// Generate a random nonce for EIP-3009 authorization
-    pub fn generate_nonce() -> H256 {
-        use rand::RngCore;
-        let mut bytes = [0u8; 32];
-        rand::thread_rng().fill_bytes(&mut bytes);
-        H256::from_slice(&bytes)
-    }
-
-    /// Verify a payment payload signature
-    pub fn verify_payment_payload(
-        payload: &crate::types::ExactEvmPayload,
-        expected_from: &str,
-        network: &str,
-    ) -> Result<bool> {
-        let from_addr = Address::from_str(expected_from)
-            .map_err(|_| X402Error::invalid_signature("Invalid from address"))?;
-
-        // Create the message hash from authorization
-        let auth = &payload.authorization;
-
-        // Get network configuration based on the payment network
-        let network_config = crate::types::NetworkConfig::from_name(network)
-            .ok_or_else(|| X402Error::invalid_signature("Unsupported network"))?;
-
-        let message_hash = eip712::create_transfer_with_authorization_hash(
-            &eip712::Domain {
-                name: "USD Coin".to_string(),
-                version: "2".to_string(),
-                chain_id: network_config.chain_id,
-                verifying_contract: Address::from_str(&network_config.usdc_contract)
-                    .map_err(|_| X402Error::invalid_signature("Invalid verifying contract"))?,
-            },
-            Address::from_str(&auth.from)
-                .map_err(|_| X402Error::invalid_signature("Invalid from address"))?,
-            Address::from_str(&auth.to)
-                .map_err(|_| X402Error::invalid_signature("Invalid to address"))?,
-            U256::from_str_radix(&auth.value, 10)
-                .map_err(|_| X402Error::invalid_signature("Invalid value"))?,
-            U256::from_str_radix(&auth.valid_after, 10)
-                .map_err(|_| X402Error::invalid_signature("Invalid valid_after"))?,
-            U256::from_str_radix(&auth.valid_before, 10)
-                .map_err(|_| X402Error::invalid_signature("Invalid valid_before"))?,
-            H256::from_str(&auth.nonce)
-                .map_err(|_| X402Error::invalid_signature("Invalid nonce"))?,
-        )?;
-
-        verify_eip712_signature(&payload.signature, message_hash, from_addr)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use ethereum_types::Address;
-
-    #[test]
-    fn test_jwt_creation() {
-        let token = jwt::create_auth_header(
-            "test_key",
-            "test_secret",
-            "api.cdp.coinbase.com",
-            "/platform/v2/x402/verify",
-        );
-        assert!(token.is_ok());
-        assert!(token.unwrap().starts_with("Bearer "));
-    }
-
-    #[test]
-    fn test_domain_creation() {
-        let domain = eip712::Domain {
-            name: "USD Coin".to_string(),
-            version: "2".to_string(),
-            chain_id: 8453,
-            verifying_contract: Address::from_str("0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913")
-                .unwrap(),
-        };
-
-        assert_eq!(domain.name, "USD Coin");
-        assert_eq!(domain.version, "2");
-        assert_eq!(domain.chain_id, 8453);
-    }
-
-    #[test]
-    fn test_nonce_generation() {
-        let nonce1 = signature::generate_nonce();
-        let nonce2 = signature::generate_nonce();
-
-        // Nonces should be different
-        assert_ne!(nonce1, nonce2);
-
-        // Nonces should be valid H256 values
-        assert_eq!(nonce1.as_bytes().len(), 32);
-        assert_eq!(nonce2.as_bytes().len(), 32);
-    }
-
-    #[test]
-    fn test_payment_payload_verification() {
-        // Create a test payment payload with valid decimal values
-        let auth = crate::types::ExactEvmPayloadAuthorization::new(
-            "0x857b06519E91e3A54538791bDbb0E22373e36b66",
-            "0x209693Bc6afc0C5328bA36FaF03C514EF312287C",
-            "1000000000000000000", // 1 USDC in wei (18 decimals)
-            "1745323800",          // Valid timestamp
-            "1745323985",          // Valid timestamp
-            "0xf3746613c2d920b5fdabc0856f2aeb2d4f88ee6037b8cc5d04a71a4462f13480", // Nonce with 0x prefix
-        );
-
-        let payload = crate::types::ExactEvmPayload {
-            signature: "0x2d6a7588d6acca505cbf0d9a4a227e0c52c6c34008c8e8986a1283259764173608a2ce6496642e377d6da8dbbf5836e9bd15092f9ecab05ded3d6293af148b571c".to_string(),
-            authorization: auth,
-        };
-
-        // This should not panic, even if verification fails
-        let result = signature::verify_payment_payload(
-            &payload,
-            "0x857b06519E91e3A54538791bDbb0E22373e36b66",
-            "base-sepolia",
-        );
-        match result {
-            Ok(_) => println!("Verification succeeded"),
-            Err(e) => println!("Verification failed with error: {}", e),
-        }
-        // The verification result might be true or false, but the function should not panic
-        // For now, we'll just check that it doesn't panic, regardless of the result
-        let _ = result;
-    }
-
-    #[test]
-    fn test_invalid_payment_payload_validation() {
-        // Test that invalid payment payloads are properly handled
-        let auth = crate::types::ExactEvmPayloadAuthorization::new(
-            "0x857b06519E91e3A54538791bDbb0E22373e36b66",
-            "0x209693Bc6afc0C5328bA36FaF03C514EF312287C",
-            "1000000",
-            "1745323800",
-            "1745323985",
-            "0xf3746613c2d920b5fdabc0856f2aeb2d4f88ee6037b8cc5d04a71a4462f13480",
-        );
-
-        let payload = crate::types::ExactEvmPayload {
-            signature: "0x2d6a7588d6acca505cbf0d9a4a227e0c52c6c34008c8e8986a1283259764173608a2ce6496642e377d6da8dbbf5836e9bd15092f9ecab05ded3d6293af148b571c".to_string(),
-            authorization: auth,
-        };
-
-        // Test with valid payload - should not panic
-        let valid_payment_payload = crate::types::PaymentPayload {
-            x402_version: 1,
-            scheme: "exact".to_string(),
-            network: "base-sepolia".to_string(),
-            payload: payload.clone(),
-        };
-
-        // This should not panic and should return a result (either Ok or Err)
-        let result = signature::verify_payment_payload(
-            &valid_payment_payload.payload,
-            "0x857b06519E91e3A54538791bDbb0E22373e36b66",
-            "base-sepolia",
-        );
-
-        // The result should be handled gracefully without panicking
-        match result {
-            Ok(_) => println!("Verification succeeded"),
-            Err(e) => println!("Verification failed with error: {}", e),
-        }
-
-        // Test that the function doesn't panic even with invalid data
-        // This test verifies that invalid data is handled gracefully
-    }
-}
